@@ -35,7 +35,7 @@ under the License.
   </scm>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>11</java.version>
         <flink.version>1.11.1</flink.version>
         <log4j.version>2.12.1</log4j.version>
         <scala.binary.version>2.11</scala.binary.version>
